+++ modules/tinymce/CHANGELOG.md
@@ -30,11 +30,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved the performance when rendering UI components #TINY-7572
 - When scrolling, the context toolbar will stick to where it was previously for large elements, such as tables #TINY-7545
 - The context toolbar will now move out of the way when it overlaps with the selection, such as in table cells #TINY-7192
+- Improved `formatter.match` to take an optional `similar` parameter #TINY-7712
 
 ### Changed
 - Changed the load order so that the content css gets loaded before the editor gets populated with contents #TINY-7249
 - Changed `emoticons`, `wordcount`, `code`, `codesample`, and `template` plugins to open dialogs using commands #TINY-7619
-- Changed `formatter.match` to take an optional `similar`-parameter #TINY-7712
 
 ### Fixed
 - The `dir` attribute was being incorrectly applied to list items #TINY-4589